cleaning up - 1
- updated overall layout
- reduced user attribute views in form and show
- link_to home on click "users" title
- cache each user on index view
- updated bulk_destroy action

+++ app/controllers/users_controller.rb
@@ -77,6 +77,7 @@ class UsersController < ApplicationController
   end
 
   def bulk_destroy
+    return unless params[:users]
     User.where(id: params[:users]).destroy_all
 
     respond_to do |format|
